Add `:nest` modifier to work around synchronization bug for objects
This commit introduces a new modifier `:nest` which nests the code fence
in a new object that's followed with a wildcard import from that object.

For example, the following markup
````md
```scala mdoc:nest
val x = 1
```
```scala mdoc
Await.result(Future(x), Duration("100ms"))
```
````

renders into the following code
```scala
object Enclosing {
  object $nest {
    val x = 1
  }
  import $nest._
  Await.result(Future(x), Duration("100ms"))
}
```

See https://stackoverflow.com/a/15176433 for an explanation why a
deadlock happens when a reference to the object `Enclosing.this`
instance leaks to a separate thread during the object's constructor.